Add Drain support to Unpacker.
Dispose() should not be block operation, so SubtreeUnpacker.Dispose() should not be blocking drain operation or some workaround should be provided.
The Drain method drains remaining items in the SubtreeUnpacker, exposes async counterpart, and is safe to duplicate call. So, calling DrainAsync(), await it, and then call Dispose() should not be blocking.