Set minClientVersion to nuspec. #177
NuGet 2.x does not support .NET Standard, so they fail to parse msgpack.nuspec file as 0.7.0.
This commit set minimum version restriction which just prevents confusing error. This fix actually does not solve the problem, it still says the client should update nuget to 2.12 or 3.x, but it gives hint to the solution.
Note: this solution is also used in corefx packages.