Remove memcmp from MessagePackString. Fix #206.
P/Invoke causes verification error in Windows Native.
It gains performance for edge case -- comparing (relatively) large blob (not string).
It is clearly rare, and it may bring packaging issue in future.
In addition, it is not so bad naive managed loop. In best (or worst) case, it only 1.8x slower than memcmp.
Although it is 12x slower when it is well aligned bytes, but it should be rare in real world.

So, memcmp has gone.