Fix build script
This commit migrates preview dotnet to latest msbuild.
Note that this script depends on Visual Studio 2017 because dotnet CLI does not
support .NET 3.5 and Xamarin yet.
In addition this script assumes that build env has path to latest msbuild since
it is hard to determine the path of msbuild due to it contains Visual Studio
edition.